# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Genetic Algorithm VRP Optimization API",
    "description": "\n# API de Otimização de Rotas com Algoritmo Genético\n\nAPI RESTful para otimização de rotas de entrega de medicamentos e suprimentos para hospitais usando Algoritmos Genéticos.\n\n## Características\n\n* **24 Operadores Genéticos**: 8 métodos de seleção, 8 operadores de crossover, 8 operadores de mutação\n* **Multi-objetivo**: Otimiza distância, prioridades, capacidade e autonomia\n* **Dados Reais**: Baseado em coordenadas de hospitais de São Paulo\n* **Execução Assíncrona**: Experimentos executam em background\n* **Persistência**: Histórico completo em SQLite\n\n## Cenários Disponíveis\n\n* `small`: ~10 hospitais, 2 veículos (testes rápidos)\n* `medium`: ~20 hospitais, 3 veículos (padrão)\n* `large`: Todos hospitais, 4-5 veículos (computacionalmente intensivo)\n* `critical`: Apenas entregas críticas/urgentes\n\n## Tecnologias\n\n* FastAPI + Uvicorn\n* SQLAlchemy + SQLite\n* NumPy + Pandas\n* Pydantic para validação\n    ",
    "contact": {
      "name": "FIAP Tech Challenge - Fase 2",
      "email": "contato@example.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "paths": {
    "/run": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Criar e executar novo experimento",
        "description": "Cria um novo experimento de otimização de rotas e inicia sua execução em background.\n\n    O experimento será executado de forma assíncrona. Use os endpoints de consulta para\n    acompanhar o progresso e obter resultados.\n\n    **Fluxo de execução:**\n    1. Valida configuração recebida\n    2. Cria registro no banco de dados (status: pending)\n    3. Inicia processamento em background\n    4. Retorna ID para consulta posterior\n\n    **Tempo estimado de execução:**\n    - Small scenario: 10-30 segundos\n    - Medium scenario: 30-90 segundos\n    - Large scenario: 2-5 minutos",
        "operationId": "run_experiment_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Experimento criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                },
                "example": {
                  "id": 42,
                  "status": "pending",
                  "message": "Experimento iniciado."
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação nos parâmetros"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
    "/experiments/{experiment_id}": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Obter detalhes de experimento específico",
        "description": "Retorna todos os detalhes de um experimento pelo seu ID.",
        "operationId": "get_experiment_experiments__experiment_id__get",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experimento encontrado",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Experimento não encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "experiments"
        ],
        "summary": "Deletar experimento específico",
        "description": "Remove um experimento específico pelo seu ID.",
        "operationId": "delete_experiment_experiments__experiment_id__delete",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experimento removido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Experimento não encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/experiments/all": {
      "delete": {
        "tags": [
          "experiments"
        ],
        "summary": "Limpar todo histórico de experimentos",
        "description": "**ATENÇÃO:** Remove TODOS os experimentos do banco de dados. Esta operação é irreversível.\n\n    Use este endpoint para limpar o histórico completo de experimentos.",
        "operationId": "delete_all_experiments_experiments_all_delete",
        "responses": {
          "200": {
            "description": "Histórico limpo com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro ao limpar histórico"
          }
        }
      }
    },
...
    "/scenarios/{scenario_name}": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "Visualizar dados do cenário",
        "description": "Retorna os dados completos de um cenário (hospitais, coordenadas, demandas, prioridades).\n\n    Útil para preview antes de executar experimento.\n\n    **Cenários disponíveis:**\n    - `small`: ~10 hospitais\n    - `medium`: ~20 hospitais\n    - `large`: Todos os hospitais (~50+)\n    - `critical`: Apenas entregas críticas/urgentes",
        "operationId": "get_scenario_preview_scenarios__scenario_name__get",
        "parameters": [
          {
            "name": "scenario_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do cenário",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Cenário inválido ou erro ao carregar"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mensagem de confirmação"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "DeleteResponse",
        "description": "Resposta de operação de deleção",
        "example": {
          "message": "Experimento 42 removido"
        }
      },
...
      "ExperimentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "ID único do experimento"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Status: pending, running, completed, failed"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mensagem descritiva"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "message"
        ],
        "title": "ExperimentResponse",
        "description": "Resposta de criação de experimento",
        "example": {
          "id": 42,
          "message": "Experimento iniciado.",
          "status": "pending"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}